Decode the upsert marker as the int4 it is

+++ apps/companion/src/github.rs
@@ -124,7 +124,7 @@ impl GithubConnector {
             outcome.events_fetched += events.len();
             for event in &events {
                 for pending in observations_for(event) {
-                    let inserted = sqlx::query_scalar::<_, i64>(
+                    let inserted = sqlx::query_scalar::<_, i32>(
                         "INSERT INTO observations (source, observed_at, kind, payload, dedupe_key) VALUES ('github', $1, $2, $3, $4) ON CONFLICT (dedupe_key) WHERE dedupe_key IS NOT NULL DO NOTHING RETURNING 1",
                     )
                     .bind(pending.observed_at)
